requirements.txt
altgraph==0.17.2
boto3==1.24.22
botocore==1.27.22
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.2
ecdsa==0.17.0
envs==1.4
future==0.18.2
idna==3.3
jmespath==1.0.1
paho-mqtt==1.6.1
pefile==2022.5.30
pyasn1==0.4.8
pycognito==2022.5.0
pycparser==2.21
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
python-dateutil==2.8.2
python-jose==3.3.0
pywin32-ctypes==0.2.0
requests==2.28.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
